Move redis logic to its own file and move frontpage renderer to a controller
